requirements.txt
aiogram==3.1.1
SQLAlchemy==2.0.23
aiosqlite==0.19.0
APScheduler==3.10.4
FastAPI==0.104.1
python-dotenv==1.0.0
pytz==2023.3.post1
uvicorn==0.24.0
